添加Address和Message的管理平台

+++ prayer-ws/src/main/resources/webroot/resources/js/users/users.js
@@ -6,7 +6,7 @@ function formUserA(path) {
     }, function () {
         FORM.init("#fUserAdd", "/static/resources/rules/validate/users/useradd.json");
         // 带不带Confirm Dialog的关闭
-        jQuery("#btnBack").attr("onclick","closeUserF()");
+        jQuery("#btnBack").attr("onclick","closeTabForm()");
     });
 }
 // Users -> User Management -> Update
@@ -19,7 +19,7 @@ function formUserU(path) {
         FORM.init("#fUserUpdate", "/static/resources/rules/validate/users/userupdate.json");
         FORM.init("#fUserPwD", "/static/resources/rules/validate/users/userpwd.json");
         // 带不带Confirm Dialog的关闭
-        jQuery("#btnBack").attr("onclick","closeUserF()");
+        jQuery("#btnBack").attr("onclick","closeTabForm()");
     });
 }
 // Users -> User Management -> View
@@ -29,7 +29,7 @@ function formUserV(path,callback) {
         "title": "User View",
         "data": data
     },function(){
-        jQuery("#btnBack").attr("onclick","closeUserV()");
+        jQuery("#btnBack").attr("onclick","closeTabView()");
         // 为删除流程量身定制按钮
         if(undefined !== callback){
             callback();
@@ -39,40 +39,28 @@ function formUserV(path,callback) {
 // Users -> User Management -> Delete
 function formUserD(){
     var data = $$U.row(".radio","#fUserListDATA");
-    BootstrapDialog.show({
-        message: "Do you want to delete this record directly ? ",
-        buttons: [{
-            label: "Yes",
-            cssClass: "btn btn-primary",
-            action: function (self) {
-                self.close();
-                _deleteUser(data);
-            }
-        },{
-            label: "No",
-            cssClass: "btn btn-danger",
-            action: function(self){
-                self.close();
-                formUserV('/dynamic/admin/forms/users/userview',function(){
-                    // 设置删除按钮
-                    jQuery("#btnDelete").removeClass("hidden");
-                });
-            }
-        }]
+    $$D.confirm("Do you want to delete this record directly ? ",function(){
+        _deleteUser(data);
+    },function(){
+        formUserV('/dynamic/admin/forms/users/userview',function(){
+            // 设置删除按钮
+            jQuery("#btnDelete").removeClass("hidden");
+        });
     });
 }
-// Users -> User Management -> Add New -> Close
-function closeUserF() {
-    TAB.closeTab(function () {
-        // 带不带Confirm Dialog的关闭
-        jQuery(".radio").attr("checked", false);
+// Users -> User Management -> Delete -> View -> Delete
+function deleteUser(){
+    $$D.confirm("Do you want to delete this record ? ",function(){
+        var data = FORM.success("#fUserView");
+        _deleteUser(data);
+        closeTabView();
     });
 }
-// Users -> User Management -> View -> Close
-function closeUserV(){
-    TAB.dismiss();
+// 执行逻辑: exeDelete
+function _deleteUser(data){
+    // TODO: 删除用户信息
+    console.log(data);
 }
-
 // Users -> User Management -> Add New -> Add New
 function addUser() {
     if ($("#fUserAdd").valid()) {
@@ -80,32 +68,6 @@ function addUser() {
         // TODO: 添加用户执行逻辑
         console.log(data);
     }
-}
-// 执行逻辑: exeDelete
-function _deleteUser(data){
-    console.log(data);
-}
-// Users -> User Management -> Delete -> View -> Delete
-function deleteUser(){
-    BootstrapDialog.show({
-        message: "Do you want to delete this record ? ",
-        buttons: [{
-            label: "Yes",
-            cssClass: "btn btn-primary",
-            action: function (self) {
-                var data = FORM.success("#fUserView");
-                _deleteUser(data);
-                self.close();
-                closeUserV();
-            }
-        },{
-            label: "No",
-            cssClass: "btn btn-danger",
-            action: function(self){
-                self.close();
-            }
-        }]
-    });
 }
 // Users -> User Management -> Update -> Update Profile
 function updateUser(){
